Fix Mastodon domain verification
Currently the 'me' rel was added in the footer vue file but that currently doesn't result in the domain being properly verified on the Mastodon profile. I ASSUME that this is because the Mastodon parser isn't running JS so can't see the tag. This change (I think?) should make it so that no JS needs to be parsed for it to read the tag, so verification should work